Move TextSizeMarkerState to separate file.
