+++ patterns/state/manipulator_state/states/selections/text_resize_state.dart
@@ -1,8 +1,6 @@
-import 'package:flutter/src/services/mouse_cursor.dart';
-
 import '../../shapes/text_shape.dart';
-import '../_/corner_marker.dart';
 import '../_/sub_states/parent_state.dart';
+import 'text_resize_marker/text_resize_marker_state.dart';
 
 class TextResizeState extends ParentState<TextShape> {
   TextResizeState({required super.selectedShape}) {
@@ -14,29 +12,5 @@ class TextResizeState extends ParentState<TextShape> {
   @override
   String toString() {
     return '${super.toString()} + Text Resize State';
-  }
-}
-
-class TextSizeMarkerState extends Marker<TextShape> {
-  TextSizeMarkerState({required super.parentState});
-
-  @override
-  void mouseDragAction(double x, double y) {
-    final newHeight = y - selectedShape.y;
-    selectedShape.resize(-1, newHeight);
-  }
-
-  @override
-  void updatePosition() {
-    final bottomCenter = selectedShape.rect.bottomCenter;
-    markerShape.move(bottomCenter.dx, bottomCenter.dy);
-  }
-
-  @override
-  MouseCursor get hoverCursor => SystemMouseCursors.resizeUpDown;
-
-  @override
-  String toString() {
-    return '${parentState.toString()} + Text Size Marker State';
   }
 }
